Refactor test vector runner from single file
