+++ tests/conformance_tests/src/stubs.rs
@@ -1,0 +1,54 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use super::*;
+
+pub struct TestRand;
+impl Rand for TestRand {
+    fn get_chain_randomness<DB: BlockStore>(
+        &self,
+        _: &DB,
+        _: DomainSeparationTag,
+        _: ChainEpoch,
+        _: &[u8],
+    ) -> Result<[u8; 32], Box<dyn StdError>> {
+        Ok(*b"i_am_random_____i_am_random_____")
+    }
+    fn get_beacon_randomness<DB: BlockStore>(
+        &self,
+        _: &DB,
+        _: DomainSeparationTag,
+        _: ChainEpoch,
+        _: &[u8],
+    ) -> Result<[u8; 32], Box<dyn StdError>> {
+        Ok(*b"i_am_random_____i_am_random_____")
+    }
+}
+
+pub struct TestSyscalls;
+impl Syscalls for TestSyscalls {
+    fn verify_signature(
+        &self,
+        _: &Signature,
+        _: &Address,
+        _: &[u8],
+    ) -> Result<(), Box<dyn StdError>> {
+        Ok(())
+    }
+    fn verify_seal(&self, _: &SealVerifyInfo) -> Result<(), Box<dyn StdError>> {
+        Ok(())
+    }
+    fn verify_post(&self, _: &WindowPoStVerifyInfo) -> Result<(), Box<dyn StdError>> {
+        Ok(())
+    }
+
+    // TODO check if this should be defaulted as well
+    fn verify_consensus_fault(
+        &self,
+        _: &[u8],
+        _: &[u8],
+        _: &[u8],
+    ) -> Result<Option<ConsensusFault>, Box<dyn StdError>> {
+        Ok(None)
+    }
+}
